WIP - env vars
