+++ cli/cmd/lib_cli_config.go
@@ -143,10 +143,6 @@ func cliEnvPromptValidation(defaults *CLIEnvConfig) *cr.PromptValidation {
 }
 
 func readTelemetryConfig() (bool, error) {
-	if os.Getenv("CORTEX_DISABLE_TELEMETRY") == "true" {
-		return false, nil
-	}
-
 	cliConfig, err := readCLIConfig()
 	if err != nil {
 		return false, err
