--- pkg/lib/k8s/aws_credentials.go
@@ -1,50 +1,0 @@
-/*
-Copyright 2019 Cortex Labs, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package k8s
-
-import (
-	kcore "k8s.io/api/core/v1"
-)
-
-func AWSCredentials() []kcore.EnvVar {
-	envVars := []kcore.EnvVar{
-		{
-			Name: "AWS_ACCESS_KEY_ID",
-			ValueFrom: &kcore.EnvVarSource{
-				SecretKeyRef: &kcore.SecretKeySelector{
-					LocalObjectReference: kcore.LocalObjectReference{
-						Name: "aws-credentials",
-					},
-					Key: "AWS_ACCESS_KEY_ID",
-				},
-			},
-		},
-		{
-			Name: "AWS_SECRET_ACCESS_KEY",
-			ValueFrom: &kcore.EnvVarSource{
-				SecretKeyRef: &kcore.SecretKeySelector{
-					LocalObjectReference: kcore.LocalObjectReference{
-						Name: "aws-credentials",
-					},
-					Key: "AWS_SECRET_ACCESS_KEY",
-				},
-			},
-		},
-	}
-
-	return envVars
-}
